2006-08-30	Joel Sherrill <joel@OARcorp.com>
	* libcsupport/include/sys/ioccom.h, libnetworking/kern/uipc_socket2.c,
	libnetworking/net/ethernet.h, libnetworking/net/if.h,
	libnetworking/net/if_ethersubr.c, libnetworking/net/if_ppp.c,
	libnetworking/net/if_pppvar.h, libnetworking/net/ppp_tty.c,
	libnetworking/net/pppcompress.c, libnetworking/net/slcompress.c,
	libnetworking/netinet/in.h, libnetworking/netinet/ip_icmp.c,
	libnetworking/netinet/tcp.h, libnetworking/netinet/tcp_seq.h,
	libnetworking/sys/protosw.h, libnetworking/sys/socketvar.h: Remove
	warnings. Most of the warnings were 16/32 bit integer sizing issues.
	Some constants had L appended, some had casts add, some types were
	changed to reflect a bit width requirement.

+++ cpukit/libcsupport/include/sys/ioccom.h
@@ -56,7 +56,7 @@
 #define	IOC_DIRMASK	0xe0000000	/* mask for IN/OUT/VOID */
 
 #define	_IOC(inout,group,num,len) \
-	((u_int32_t)inout | \
+	(u_int32_t) ((u_int32_t)inout | \
          (u_int32_t) ((u_int32_t)((u_int32_t)len & IOCPARM_MASK) << 16) | \
          (u_int32_t)((group) << 8) | \
          (u_int32_t)(num))
