+++ cpukit/libnetworking/netinet/in.h
@@ -299,7 +299,7 @@ struct sockaddr_in {
 #define	INADDR_ALLRTRS_GROUP	(u_int32_t)0xe0000002	/* 224.0.0.2 */
 #define	INADDR_MAX_LOCAL_GROUP	(u_int32_t)0xe00000ff	/* 224.0.0.255 */
 
-#define	IN_LOOPBACKNET		127			/* official! */
+#define	IN_LOOPBACKNET		(u_int32_t)127			/* official! */
 
 /*
  * Structure used to describe IP options.
